Ava2024/be/requirements.txt
Django==3.2.5
psycopg2-binary==2.9.3,<3.0
psycopg2>=2.9
djangorestframework==3.12.4
django-cors-headers==3.13.0
gunicorn==20.1.0